// src/consts/enums/ids/skill-ids.enum.ts
export enum SkillIds {
  // Exploration Skill Tree
  autoWaveProgression = 0,
  haste = 1,
  goldGain = 2,
  knowledge = 3,
  weaknesses = 4,

  // Damage Skill Tree
  attackPower = 205,
  attackSpeed = 206,
  overkill = 207,
  critChance = 208,
  critMulti = 209,
  doubleAttack = 210,

  // Magic Skill Tree
  fireStrike = 400,
  maxMana = 401,
  manaRegen = 402,
  magicDamage = 403,
  spellCrit = 404,
  spellCooldown = 405,
}